@@ -39,44 +39,37 @@ dependencies:
   cupertino_icons: ^1.0.5
   qr_flutter: ^4.1.0
   qr_code_scanner: ^1.0.1
-  # http: ^1.1.0
-  # http_parser: ^4.0.0
-  shared_preferences: ^2.1.0
-  url_launcher: ^6.1.10
-  # flutter_rounded_date_picker: ^2.0.2
-  checkbox_formfield: ^0.2.0
+  shared_preferences: ^2.3.1
+  url_launcher: ^6.3.0
   share_files_and_screenshot_widgets: ^1.0.6
   image_gallery_saver: ^2.0.3
-  connectivity_plus: ^5.0.2
-  flutter_launcher_icons: ^0.13.0
-  flutter_native_splash: ^2.2.19
-  # flutter_blue: ^0.8.0
-  geolocator: ^10.1.0
-  # location: ^4.4.0
-  # app_settings: ^4.2.0
+  connectivity_plus: ^6.0.4
+  flutter_launcher_icons: ^0.13.1
+  flutter_native_splash: ^2.4.1
+  geolocator: ^12.0.0
   flutter_phone_direct_caller: ^2.1.1
-  permission_handler: ^11.0.1
-  image_picker: ^1.0.5
+  permission_handler: ^11.3.1
+  image_picker: ^1.1.2
   flutter_swiper_plus: ^2.0.4
-  intl: ^0.18.0
-  dio: ^5.3.3
-  get: ^4.6.5
-  restart_app: ^1.1.2
+  intl: ^0.19.0
+  dio: ^5.5.0
+  get: ^4.6.6
+  restart_app: ^1.2.1
   flutter_easyloading: ^3.0.5
-  google_maps_flutter: ^2.5.3
-  easy_loading_button: ^0.3.0
-  calendar_date_picker2: ^0.5.3
-  animated_custom_dropdown: ^2.0.0
-  flutter_local_notifications: ^16.2.0
+  google_maps_flutter: ^2.9.0
+  easy_loading_button: ^0.3.2
+  calendar_date_picker2: ^1.1.7
+  animated_custom_dropdown: ^3.1.1
+  flutter_local_notifications: ^17.2.1
   flutter_phoenix: ^1.1.1
-  flutter_datetime_picker_bdaya: ^3.0.1
-  package_info_plus: ^4.1.0
-  dotted_line: ^3.2.0
-  local_auth: ^2.1.6
-  firebase_messaging: ^14.7.16
-  firebase_core: ^2.25.4
-  # google_mlkit_text_recognition: ^0.10.0
-  camera: ^0.10.5+5
+  package_info_plus: ^8.0.1
+  dotted_line: ^3.2.2
+  local_auth: ^2.3.0
+  firebase_messaging: ^15.1.2
+  firebase_core: ^3.5.0
+  camera: ^0.11.0+2
+  slide_digital_clock: ^1.0.3
+  image: ^4.2.0
 
   
   
