@@ -6,7 +6,7 @@ import Firebase
 import FirebaseMessaging
 import flutter_local_notifications
 
-@UIApplicationMain
+@main
 @objc class AppDelegate: FlutterAppDelegate {
   override func application(
     _ application: UIApplication,
