@@ -4,18 +4,18 @@ import 'package:doormster/style/styleButton/ButtonStyle.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-void dialogTwobutton_Subtitle(
-  title,
-  subtitle,
-  icon,
-  coloricon,
-  button1,
-  press1,
-  button2,
-  press2,
-  click,
-  willpop,
-) {
+void dialogTwobutton_Subtitle({
+  required title,
+  required subtitle,
+  required icon,
+  required colorIcon,
+  required textButton1,
+  required press1,
+  required textButton2,
+  required press2,
+  required click,
+  required willpop,
+}) {
   Get.dialog(
       transitionDuration: Duration.zero,
       barrierDismissible: click,
@@ -32,7 +32,7 @@ void dialogTwobutton_Subtitle(
                   Icon(
                     icon,
                     size: 60,
-                    color: coloricon,
+                    color: colorIcon,
                   ),
                   Text(
                     title,
@@ -64,7 +64,7 @@ void dialogTwobutton_Subtitle(
                                 Get.theme.primaryColor,
                                 BorderRadius.circular(5)),
                             child: Text(
-                              button1,
+                              textButton1,
                               style: TextStyle(
                                   fontSize: 16,
                                   letterSpacing: 1,
@@ -83,7 +83,7 @@ void dialogTwobutton_Subtitle(
                                 Get.theme.primaryColor,
                                 BorderRadius.circular(5)),
                             child: Text(
-                              button2,
+                              textButton2,
                               style: TextStyle(
                                   fontSize: 16,
                                   letterSpacing: 1,
