@@ -1,13 +1,11 @@
 import 'dart:async';
 import 'package:doormster/components/bottombar/bottom_controller.dart';
-import 'package:doormster/components/bottombar/bottombar.dart';
-import 'package:doormster/controller/get_info.dart';
+import 'package:doormster/components/bottombar/navigation_ids.dart';
+import 'package:doormster/controller/branch_controller.dart';
 import 'package:doormster/controller/home_controller.dart';
+import 'package:doormster/controller/news_controller.dart';
 import 'package:doormster/controller/profile_controller.dart';
-import 'package:doormster/screen/main_screen/home_page.dart';
-import 'package:doormster/screen/main_screen/news_page.dart';
 import 'package:doormster/screen/main_screen/notification_page.dart';
-import 'package:doormster/screen/main_screen/profile_page.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
@@ -18,28 +16,30 @@ class NavbarNotifier extends ChangeNotifier {
     bool exitingApp = true;
     switch (index) {
       case 0:
-        if (homeKey.currentState != null && homeKey.currentState!.canPop()) {
-          homeKey.currentState!.pop();
+        if (Keys.home?.currentState != null &&
+            Keys.home!.currentState!.canPop()) {
+          Keys.home?.currentState!.pop();
           exitingApp = false;
         }
         break;
       case 1:
-        if (newsKey.currentState != null && newsKey.currentState!.canPop()) {
-          newsKey.currentState!.pop();
+        if (Keys.news?.currentState != null &&
+            Keys.news!.currentState!.canPop()) {
+          Keys.news?.currentState!.pop();
           exitingApp = false;
         }
         break;
       case 2:
-        if (notifyKey.currentState != null &&
-            notifyKey.currentState!.canPop()) {
-          notifyKey.currentState!.pop();
+        if (Keys.notify?.currentState != null &&
+            Keys.notify!.currentState!.canPop()) {
+          Keys.notify?.currentState!.pop();
           exitingApp = false;
         }
         break;
       case 3:
-        if (profileKey.currentState != null &&
-            profileKey.currentState!.canPop()) {
-          profileKey.currentState!.pop();
+        if (Keys.profile?.currentState != null &&
+            Keys.profile!.currentState!.canPop()) {
+          Keys.profile?.currentState!.pop();
           exitingApp = false;
         }
         break;
@@ -56,29 +56,27 @@ class NavbarNotifier extends ChangeNotifier {
   void popAllRoutes(int index) async {
     switch (index) {
       case 0:
-        if (homeKey.currentState != null && homeKey.currentState!.canPop()) {
-          homeKey.currentState?.popUntil((route) => route.isFirst);
+        if (Keys.home?.currentState != null &&
+            Keys.home!.currentState!.canPop()) {
+          Keys.home?.currentState?.popUntil((route) => route.isFirst);
         } else if (bottomController.selectedIndex.value == 0) {
-          get_Info().then((value) => Homecontroller.GetMenu());
+          Homecontroller.GetMenu();
         }
         return;
       case 1:
-        if (newsKey.currentState != null && newsKey.currentState!.canPop()) {
-          newsKey.currentState!.popUntil((route) => route.isFirst);
+        if (Keys.news?.currentState != null &&
+            Keys.news!.currentState!.canPop()) {
+          Keys.news?.currentState!.popUntil((route) => route.isFirst);
         } else if (bottomController.selectedIndex.value == 1) {
-          newsKey.currentState?.pushReplacement(
-            GetPageRoute(
-                page: () => const News_Page(),
-                transitionDuration: Duration.zero),
-          );
+          news_controller.get_News(branch_id: branchController.branch_Id.value);
         }
         return;
       case 2:
-        if (notifyKey.currentState != null &&
-            notifyKey.currentState!.canPop()) {
-          notifyKey.currentState!.popUntil((route) => route.isFirst);
+        if (Keys.notify?.currentState != null &&
+            Keys.notify!.currentState!.canPop()) {
+          Keys.notify?.currentState!.popUntil((route) => route.isFirst);
         } else if (bottomController.selectedIndex.value == 2) {
-          notifyKey.currentState?.pushReplacement(
+          Keys.notify?.currentState?.pushReplacement(
             GetPageRoute(
                 page: () => const Notification_Page(),
                 transitionDuration: Duration.zero),
@@ -86,11 +84,11 @@ class NavbarNotifier extends ChangeNotifier {
         }
         return;
       case 3:
-        if (profileKey.currentState != null &&
-            profileKey.currentState!.canPop()) {
-          profileKey.currentState!.popUntil((route) => route.isFirst);
+        if (Keys.profile?.currentState != null &&
+            Keys.profile!.currentState!.canPop()) {
+          Keys.profile?.currentState!.popUntil((route) => route.isFirst);
         } else if (bottomController.selectedIndex.value == 3) {
-          Profilecontroller.getInfo(500);
+          profileController.get_Profile(loadingTime: 100);
         }
         return;
       default:
