@@ -11,27 +11,24 @@ class Logo_Opacity extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Center(
-      child: Container(
-          color: Theme.of(context).scaffoldBackgroundColor,
-          child: Obx(
-            () => Opacity(
-              opacity: 0.5,
-              child: Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 15),
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    themeController.imageTheme(scale: 5.0),
-                    Text(
-                      title,
-                      textAlign: TextAlign.center,
-                      style: TextStyle(fontSize: 18),
-                    )
-                  ],
-                ),
-              ),
-            ),
-          )),
-    );
+        child: Obx(
+      () => Opacity(
+        opacity: 0.5,
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 15),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              themeController.imageTheme(scale: 5.0),
+              Text(
+                title,
+                textAlign: TextAlign.center,
+                style: TextStyle(fontSize: 18),
+              )
+            ],
+          ),
+        ),
+      ),
+    ));
   }
 }
