@@ -1,67 +1,58 @@
-import 'dart:developer';
-
-import 'package:dio/dio.dart';
-import 'package:doormster/components/actions/disconnected_dialog.dart';
-import 'package:doormster/components/alertDialog/alert_dialog_onebutton_subtext.dart';
-import 'package:doormster/components/snackbar/snackbar.dart';
 import 'package:doormster/controller/login_controller.dart';
 import 'package:doormster/screen/main_screen/login_page.dart';
 import 'package:doormster/screen/main_screen/login_staff_page.dart';
-import 'package:doormster/service/connected/connect_api.dart';
 import 'package:doormster/service/notify/notify_token.dart';
-import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-final LogoutController logoutController = Get.put(LogoutController());
+final LogoutController logoutController = LogoutController();
 
-class LogoutController extends GetxController {
+class LogoutController {
   Future<void> logout() async {
-    final prefs = await SharedPreferences.getInstance();
-    final security = prefs.getBool('security');
-    //ลบ token device notify
-    await Notify_Token().deletenotifyToken();
-    //เรียก keys ที่แชร์ไว้มารวมกัน ใน allKeys
-    Set<String> allKeys = prefs.getKeys();
+    try {
+      final prefs = await SharedPreferences.getInstance();
 
-    //loop keys เพื่อหา key ที่ไม่ต้องการ remove
-    // for (String key in allKeys) {
-    //   if (key != 'notifyToken' &&
-    //       key != 'security' &&
-    //       key != 'language' &&
-    //       key != 'theme') {
-    //     prefs.remove(key);
-    //   }
-    // }
+      //ลบ token device notify
+      // await Notify_Token().deletenotifyToken();
+      //เรียก keys ที่แชร์ไว้มารวมกัน ใน allKeys
+      Set<String> allKeys = prefs.getKeys();
 
-    // if (security == true) {
-    //   Get.offAll(() => Login_Staff());
-    // } else {
-    //   Get.offAll(() => Login_Page());
-    // }
-
-    if (security == true) {
+      //loop keys เพื่อหา key ที่ไม่ต้องการ remove
       for (String key in allKeys) {
-        if (key != 'notifyToken' && key != 'language' && key != 'theme') {
+        if (key != 'username' &&
+            key != 'password' &&
+            key != 'language' &&
+            key != 'remember' &&
+            key != 'theme') {
           prefs.remove(key);
         }
       }
-      log(allKeys.toString());
-      Get.offAll(() => Login_Staff());
-    } else {
-      prefs.remove('token');
-      prefs.remove('role');
-      prefs.remove('deviceId');
-      prefs.remove('weiganId');
-      prefs.remove('security');
-      prefs.remove('image');
-      prefs.remove('fname');
-      prefs.remove('lname');
-      log(allKeys.toString());
       Get.offAll(() => Login_Page());
+
+      // if (security == true) {
+      //   for (String key in allKeys) {
+      //     if (key != 'notifyToken' &&
+      //         key != 'language' &&
+      //         key != 'theme' &&
+      //         key != 'remember' &&
+      //         key != 'security') {
+      //       prefs.remove(key);
+      //     }
+      //   }
+      //   Get.offAll(() => Login_Staff());
+      // } else {
+      //   prefs.remove('token');
+      //   prefs.remove('deviceId');
+      //   prefs.remove('weiganId');
+      //   prefs.remove('role');
+      //   prefs.remove('image');
+      //   prefs.remove('fname');
+      //   prefs.remove('lname');
+      //   Get.offAll(() => Login_Page());
+      // }
+    } catch (e) {
+      print('Logout failed: $e');
+      Get.snackbar('Error', 'Failed to logout. Please try again.');
     }
-    loginController.loadUsernamePassword();
-    print('logout success');
   }
 }
