@@ -9,22 +9,30 @@ const Map<String, String> en_US = {
   'remember': 'Remember me',
   'user': 'User',
   'employee': 'Employee',
-  'no_account': 'Not a member of HIP Smart Community yet ',
+  'no_account':
+      "If you don't have an account, please contact the administrator.",
   'login_for': 'Login for ',
   'login_for_employee': 'Login for employee',
 
   //Register page
   'register_title': 'Register',
   'username': 'Username',
+  'idCard': 'ID card number',
   'fname': 'Name',
   'lname': 'Lastname',
   'email': 'Email',
-  'company': 'Select Company',
+  'sex': 'Sex',
+  'male': 'Male',
+  'female': 'Female',
+  'day': 'Day',
+  'month': 'Month',
+  'year': 'Year(Born)',
+  'company': 'Select company',
   'password': 'Password',
-  'current_password': 'Current Password',
-  'new_password': 'New Password',
-  'confirm_password': 'Confirm Password',
-  'accept_terms': 'I accept the Terms of Service.',
+  'current_password': 'Current password',
+  'new_password': 'New password',
+  'confirm_password': 'Confirm password',
+  'accept_terms': 'I accept the terms of service.',
 
   //Bottom bar
   'home': 'Home',
@@ -36,7 +44,7 @@ const Map<String, String> en_US = {
   'info': 'Personal Info',
   'edit_profile': 'Edit Profile',
   'setting': 'Settings',
-  'switch_company': 'Switch Company',
+  'select_village': 'Select a village',
   'add_company': 'Add New Company',
 
   //Qr smart
@@ -47,37 +55,40 @@ const Map<String, String> en_US = {
   'emergancy_call': 'Emergancy Call',
 
   //Register visitor
-  'visitor_info': 'Visitor Information',
-  'create_visitor': 'Register Visitor',
-  'visitor_name': 'Visitor Name',
-  'contacts': 'Contacts',
-  'fullname': 'Full Name',
-  'phone': 'Contact Number',
-  'phone_number': 'Phone Number',
+  'visitor_info': 'Visitor information',
+  'create_visitor': 'Register visitor',
+  'visitor_name': 'Visitor name',
+  'meet_name': 'Meet name',
+  'fullname': 'Full name',
+  'phone': 'Contact number',
+  'phone_number': 'Phone number',
   'device_type': 'Device Type',
-  'device': 'Selact Device',
-  'no_device': 'No Device',
-  'access': 'Access Rights',
-  'pick_date': 'Select Date',
-  'start': 'Start',
-  'end': 'End',
-  'access_count': 'Access Count',
+  'device': 'Selact device',
+  'no_device': 'No device',
+  'access': 'Access rights',
+  'pick_date': 'Select date',
+  'choose_place': 'Choose a place.',
+  'start_date': 'Start date',
+  'end_date': 'End date',
+  'access_count': 'Access count',
   'count': 'Count',
   'create_qrcode': 'Create QRCode',
+  'house_number': 'House number',
+  'place_meet': 'Place to meet',
 
   //Security guard
-  'checkIn': 'Check-In',
+  'checkIn': 'Check in',
   'register_point': 'Register Checkpoint',
-  'report': 'View Check Reports',
-  'checkpoint_report': 'Checkpoint Reports',
-  'date_report': 'Date Report',
-  'check_extra_round': 'Check Extra Round',
-  'extra_round': 'Extra Round',
-  'view_round': 'View Check Rounds',
-  'view_checkpoint': 'View Checkpoints',
-  'by_point': 'Check by Points',
-  'extra_checkpoint': 'Check Extra Point',
-  'extra_point': 'Extra Point',
+  'report': 'View check reports',
+  'checkpoint_report': 'Checkpoint reports',
+  'date_report': 'Date report',
+  'check_extra_round': 'Check extra round',
+  'extra_round': 'Extra round',
+  'view_round': 'View check rounds',
+  'view_checkpoint': 'View checkpoints',
+  'by_point': 'Check by points',
+  'extra_checkpoint': 'Check extra point',
+  'extra_point': 'Extra point',
   'emergency': 'Emergency',
   'round': 'Round',
   'interval': 'Interval',
@@ -85,59 +96,69 @@ const Map<String, String> en_US = {
   'record': 'There are',
   'no_record': 'There are no inspection records.',
   'checkpoint': 'Checkpoint',
-  'checkpoint_name': 'Checkpoint Name',
+  'checkpoint_name': 'Checkpoint name',
   'checklist': 'Checklist',
-  'add_checklist': 'Add Checklist',
+  'add_checklist': 'Add checklist',
   'checkpoint_location': 'Location',
-  'view_location': 'View Location',
-  'current position': 'Current Position',
-  'status': 'Status : ',
+  'view_location': 'View location',
+  'current_position': 'Current position',
+  'status': 'Status',
   'date': 'Date',
   'time': 'Time',
-  'record_time': 'Record Time',
+  'record_time': 'Record time',
   'event': 'Event',
-  'event_record': 'Event Record',
-  'select_event': 'Select Event',
-  'desciption': 'Desciption',
-  'illustration': 'Illustration',
+  'event_record': 'Event record',
+  'select_event': 'Select event',
+  'normal': 'Normal',
+  'abnormal': 'Abnormal',
+  'description': 'Description',
+  'add_image': 'Add images',
   'pictures': 'Pictures',
-  'view_pictures': 'View Pictures',
+  'view_pictures': 'View pictures',
   'no_regis': 'Not yet registered',
   'registered': 'Registered',
   'no_regis_checkpoint': 'This checkpoint has not been registered yet.',
-  'checked_complete': 'Checked Complete',
-  'not_checked': 'Not Checked Yet',
-  'checked_incomplete': 'Checked Incomplete',
+  'checked_complete': 'Checked complete',
+  'not_checked': 'Not checked yet',
+  'checked_incomplete': 'Checked incomplete',
   'need_check': 'Rounds that need to be checked.',
 
   //Setting
-  'change_password': 'Change Password',
-  'change_language': 'Change Language',
-  'select_language': 'Select Language',
-  'change_theme': 'Change Theme',
-  'select_theme': 'Select Theme',
+  'change_password': 'Change password',
+  'change_language': 'Change language',
+  'select_language': 'Select language',
+  'change_theme': 'Change theme',
+  'select_theme': 'Select theme',
   'light': 'Light',
   'dark': 'Dark',
   'system': 'System',
-  'disable_account': 'Disable Account',
+  'disable_account': 'Disable account',
   'version': 'Version',
 
   //Profile
-  'change_picture': 'Change Profile Picture',
+  'change_picture': 'Change profile picture',
   'camera': 'Camera',
-  'photo': 'Select Photo',
+  'photo': 'Select photo',
+
+  //Complaint
+  'complaint': 'Complaint',
+  'complaint_list': 'Complaint list',
+  'title': 'Title',
+  'type': 'Type',
+  'type_complaint': 'Complaint type',
+  'resolved': '*If the issue has been resolved.',
 
   //Search
   'search': 'Search',
-  'search_door': 'Search Door',
-  'search_round': 'Search Round',
-  'search_checkpoint': 'Search Checkpoint',
+  'search_door': 'Search door',
+  'search_round': 'Search round',
+  'search_checkpoint': 'Search checkpoint',
 
   //Title appbar
-  'news_feed': 'News Feed',
-  'list_checkpoint': 'Checkpoint List',
-  'list_report': 'Checklist Reports',
-  'list_round': 'Checklist Rounds',
+  'news_feed': 'News feed',
+  'list_checkpoint': 'Checkpoint list',
+  'list_report': 'Checklist reports',
+  'list_round': 'Checklist rounds',
 
   //Opacity title
   'contact_admin_approve':
@@ -147,13 +168,14 @@ const Map<String, String> en_US = {
   'contact_admin_manu':
       'There are no menus accessible for you. Please contact the administrator.',
   'check_connect': 'Please check the connection.',
-  'door_not_found': 'Door Not Found',
-  'device_not_found': 'Device Not Found',
-  'data_not_found': 'Data Not Found',
-  'no_news': 'No News Yet',
-  'no_notification': 'No Notification',
-  'no_data': 'No Recorded Data',
-  'no_picture': 'No Picture',
+  'door_not_found': 'Door not found',
+  'device_not_found': 'Device not found',
+  'data_not_found': 'Data not found',
+  'no_news': 'No news yet',
+  'no_notification': 'No notification',
+  'no_data': 'No recorded data',
+  'no_picture': 'No picture',
+  'image_error': 'Unable to load image.',
 
   //Button
   'ok': 'Done',
@@ -162,9 +184,10 @@ const Map<String, String> en_US = {
   'submit': 'Submit',
   'save': 'Save',
   'cancel': 'Cancel',
+  'close': 'Closed',
   'share': 'Share',
-  'see_more': 'See More',
-  'see_less': 'See Less',
+  'see_more': 'See more',
+  'see_less': 'See less',
   'today': 'Today',
   'all': 'All',
   'to': 'to',
@@ -178,8 +201,12 @@ const Map<String, String> en_US = {
   'register_fail_again': 'Registration failed.  Please try again.',
   'press_again': 'Press again to exit.',
 
-  'found_error': 'Error found.',
-  'connect_fail': 'Unable to connect. Please try again.',
+  'occur_error': 'Error occurred.',
+  'connect_error': 'Unable to connect. Please try again.',
+  'error_timeout': 'Connection timeout.',
+  'connect_timeout':
+      'Unable to connect within the time limit. Please try again.',
+  'invalid_data': 'Invalid data. Please try again.',
   'connect_internet_pls': 'Please connect to the internet.',
   'again_pls': 'Please try again.',
   'account_block': 'This account has been suspended.',
@@ -213,12 +240,17 @@ const Map<String, String> en_US = {
 
   'checkpoint_regis': 'This checkpoint is already registered.',
   'checkpoint_no_regis': 'This checkpoint is not registered.',
-  'checkin_success': 'Check In successfully.',
+  'checkin_success': 'Check in successfully.',
   'checkpoint_found': 'Checkpoint not found.',
+  'checkpoint_found_round': 'This checkpoint was not found in this round.',
   'repeat_checkpoint': 'Repeat checkpoint.',
   'data_checkpoint': 'Data for this checkpoint has already been saved.',
   'save_success': 'Save successful.',
   'save_unsuccess': 'Save unsuccessful. Please try again.',
+  'cancel_success': 'Cancel successful.',
+  'cancel_unsuccess': 'Cancel unsuccessful. Please try again.',
+  'close_success': 'Close successful.',
+  'close_unsuccess': 'Close unsuccessful. Please try again.',
   'check_later':
       "It's not yet time for your check round, please try again later.",
   'create_qrcode_success': 'Create QR Code successfully.',
@@ -234,8 +266,11 @@ const Map<String, String> en_US = {
   'capture_success': 'Image save successful.',
   'capture_fail': 'Image save failed.',
   'capture_fail_again': 'Image save failed, please try again.',
+  'wrong_date': 'Invalid date',
+  'end_start': 'End date must not be earlier than the start date',
 
   // Error text form
+  'enter_title': 'Please enter a title.',
   'enter_info': 'Enter Info',
   'enter_info_pls': 'Please enter info.',
   'enter_info_error': 'Incomplete information entered.',
@@ -259,10 +294,26 @@ const Map<String, String> en_US = {
   'phone_10char': 'Please enter a 10-digit phone number.',
   'select_type': 'Please select a type.',
   'select_date': 'Please select a date.',
+  'select_place': 'Please select a place.',
   'enter_access_count': 'Please enter the access count.',
 
   'enter_checklist': 'Enter checklist.',
   'select_checklist': 'Please select a checklist item.',
   'select_event_pls': 'Please select an event.',
-  'enter_desciption': 'Abnormal event, please add details.',
+  'ab_enter_description': 'Abnormal event, please add description.',
+  'enter_description': 'Please add description.',
+
+  //Complaint staus
+  '1cp': 'Pending receipt',
+  '2cp': 'Complaint received',
+  '3cp': 'Under review',
+  '4cp': 'Resolved',
+  '5cp': 'Closed',
+  '6cp': 'Cancelled',
+
+  //SOS staus
+  '1sos': 'Pending receipt',
+  '2sos': 'Under review',
+  '3sos': 'Resolved',
+  '4sos': 'Cancelled',
 };
