@@ -9,16 +9,23 @@ const Map<String, String> th_TH = {
   'remember': 'บันทึกรหัสผ่าน',
   'user': 'ผู้ใช้ทั่วไป',
   'employee': 'พนักงาน',
-  'no_account': 'ยังไม่ได้เป็นสมาชิก HIP Smart Community ',
+  'no_account': 'หากยังไม่มีบัญชี ให้ติดต่อผู้ดูแล',
   'login_for': 'เข้าสู่ระบบใช้งาน ',
   'login_for_employee': 'เข้าสู่ระบบใช้งาน สำหรับพนักงาน',
 
   //Register page
   'register_title': 'ลงทะเบียน',
   'username': 'ชื่อผู้ใช้',
+  'idCard': 'เลขบัตรประชาชน',
   'fname': 'ชื่อ',
   'lname': 'นามสกุล',
   'email': 'อีเมล',
+  'sex': 'เพศ',
+  'male': 'ชาย',
+  'female': 'หญิง',
+  'day': 'วัน',
+  'month': 'เดือน',
+  'year': 'ปี(เกิด)',
   'company': 'เลือกบริษัท',
   'password': 'รหัสผ่าน',
   'current_password': 'รหัสผ่านปัจจุบัน',
@@ -36,7 +43,7 @@ const Map<String, String> th_TH = {
   'info': 'ข้อมูลส่วนตัว',
   'edit_profile': 'แก้ไขข้อมูลส่วนตัว',
   'setting': 'การตั้งค่า',
-  'switch_company': 'สลับโครงการ',
+  'select_village': 'เลือกหมู่บ้าน',
   'add_company': 'เพิ่มโครงการใหม่',
 
   //Qr smart
@@ -50,20 +57,23 @@ const Map<String, String> th_TH = {
   'visitor_info': 'ข้อมูลผู้มาติดต่อ',
   'create_visitor': 'ลงทะเบียนผู้มาติดต่อ',
   'visitor_name': 'ชื่อผู้มาติดต่อ',
-  'contacts': 'ติดต่อพบ',
+  'meet_name': 'ชื่อผู้ที่ต้องการพบ',
   'fullname': 'ชื่อ - สกุล',
   'phone': 'เบอร์ติดต่อ',
-  'phone_number': 'เบอร์โทร',
+  'phone_number': 'เบอร์โทรศัพท์',
   'device_type': 'ประเภทอุปกรณ์',
   'device': 'เลือกอุปกรณ์',
   'no_device': 'ไม่มีอุปกรณ์',
   'access': 'สิทธิ์การเข้าถึง',
   'pick_date': 'เลือกวันที่',
-  'start': 'เริ่มต้น',
-  'end': 'สิ้นสุด',
+  'choose_place': 'เลือกสถานที่',
+  'start_date': 'วันที่เริ่มต้น',
+  'end_date': 'วันที่สิ้นสุด',
   'access_count': 'สิทธิ์การใช้งาน',
   'count': 'จำนวนครั้ง',
   'create_qrcode': 'สร้าง QRCode',
+  'house_number': 'บ้านเลขที่',
+  'place_meet': 'สถานที่ต้องการพบ',
 
   //Security guard
   'checkIn': 'ตรวจจุด',
@@ -91,16 +101,18 @@ const Map<String, String> th_TH = {
   'checkpoint_location': 'ตำแหน่งจุดตรวจ',
   'view_location': 'ดูตำแหน่ง',
   'current_position': 'ตำแหน่งปัจจุบัน',
-  'status': 'สถานะ : ',
+  'status': 'สถานะ',
   'date': 'วันที่',
   'time': 'เวลา',
   'record_time': 'เวลาที่บันทึก',
   'event': 'เหตุการณ์',
   'event_record': 'บันทึกเหตุการณ์',
   'select_event': 'เลือกเหตุการณ์',
-  'desciption': 'รายละเอียดเพิ่มเติม',
-  'illustration': 'รูปภาพประกอบ',
-  'pictures': 'รูปภาพการตรวจ',
+  'normal': 'ปกติ',
+  'abnormal': 'ไม่ปกติ',
+  'description': 'รายละเอียดเพิ่มเติม',
+  'add_image': 'เพิ่มรูปภาพ',
+  'pictures': 'รูปภาพประกอบ',
   'view_pictures': 'ดูรูปภาพ',
   'no_regis': 'ยังไม่ลงทะเบียน',
   'registered': 'ลงทะเบียนแล้ว',
@@ -127,6 +139,14 @@ const Map<String, String> th_TH = {
   'camera': 'ถ่ายรูป',
   'photo': 'เลือกรูปภาพ',
 
+  //Complaint
+  'complaint': 'ร้องเรียน',
+  'complaint_list': 'รายการร้องเรียน',
+  'title': 'หัวข้อ',
+  'type': 'ประเภท',
+  'type_complaint': 'ประเภทคำร้องเรียน',
+  'resolved': '*หากปัญหาถูกแก้ไขแล้ว',
+
   //Search
   'search': 'ค้นหา',
   'search_door': 'ค้นหาประตู',
@@ -151,6 +171,7 @@ const Map<String, String> th_TH = {
   'no_notification': 'ไม่มีการแจ้งเตือน',
   'no_data': 'ไม่มีข้อมูลที่บันทึก',
   'no_picture': 'ไม่มีรูปภาพ',
+  'image_error': 'ไม่มีสามารถโหลดรูปภาพได้',
 
   //Button
   'ok': 'ตกลง',
@@ -159,6 +180,7 @@ const Map<String, String> th_TH = {
   'submit': 'ยืนยัน',
   'save': 'บันทึก',
   'cancel': 'ยกเลิก',
+  'close': 'ปิดคำร้อง',
   'share': 'แชร์',
   'see_more': 'ดูเพิ่มเติม',
   'see_less': 'ดูน้อยลง',
@@ -175,8 +197,11 @@ const Map<String, String> th_TH = {
   'register_fail_again': 'ลงทะเบียนไม่สำเร็จ กรุณาลองใหม่อีกครั้ง',
   'press_again': 'กดอีกครั้งเพื่อออก',
 
-  'found_error': 'พบข้อผิดพลาด',
-  'connect_fail': 'ไม่สามารถเชื่อมต่อได้ กรุณาลองใหม่อีกครั้ง',
+  'occur_error': 'เกิดข้อผิดพลาด',
+  'connect_error': 'ไม่สามารถเชื่อมต่อได้ กรุณาลองใหม่อีกครั้ง',
+  'error_timeout': 'การเชื่อมต่อหมดเวลา',
+  'connect_timeout': 'ไม่สามารถเชื่อมต่อได้ในเวลาที่กำหนด กรุณาลองใหม่อีกครั้ง',
+  'invalid_data': 'ข้อมูลไม่ถูกต้อง กรุณาลองใหม่อีกครั้ง',
   'connect_internet_pls': 'กรุณาเชื่อมต่ออินเตอร์เน็ต',
   'again_pls': 'กรุณาลองใหม่อีกครั้ง',
   'account_block': 'บัญชีนี้ถูกระงับการใช้งาน',
@@ -210,10 +235,15 @@ const Map<String, String> th_TH = {
   'checkpoint_no_regis': 'จุดตรวจนี้ ยังไม่ได้ลงทะเบียน',
   'checkin_success': 'ตรวจเช็คสำเร็จ',
   'checkpoint_found': 'ไม่พบจุดตรวจ',
+  'checkpoint_found_round': 'ไม่พบจุดตรวจนี้ ในรอบนี้',
   'repeat_checkpoint': 'ตรวจจุดซ้ำ',
   'data_checkpoint': 'จุดตรวจนี้ ถูกบันทึกข้อมูลแล้ว',
   'save_success': 'บันทึกสำเร็จ',
   'save_unsuccess': 'บันทึกไม่สำเร็จ กรุณาลองใหม่อีกครั้ง',
+  'cancel_success': 'ยกเลิกสำเร็จ',
+  'cancel_unsuccess': 'ยกเลิกไม่สำเร็จ กรุณาลองใหม่อีกครั้ง',
+  'close_success': 'ปิดคำร้องสำเร็จ',
+  'close_unsuccess': 'ปิดคำร้องไม่สำเร็จ กรุณาลองใหม่อีกครั้ง',
   'check_later': 'ยังไม่ถึงรอบเดินตรวจ โปรดลองใหม่ในภายหลัง',
   'create_qrcode_success': 'สร้าง QR Code สำเร็จ',
   'create_qrcode_fail': 'สร้าง QR Code ไม่สำเร็จ',
@@ -227,8 +257,11 @@ const Map<String, String> th_TH = {
   'capture_success': 'บันทึกภาพสำเร็จ',
   'capture_fail': 'บันทึกภาพไม่สำเร็จ',
   'capture_fail_again': 'บันทึกภาพไม่สำเร็จ กรุณาลองใหม่อีกครั้ง',
+  'wrong_date': 'วันที่ไม่ถูกต้อง',
+  'end_start': 'วันที่สิ้นสุด ต้องไม่น้อยกว่า วันที่เริ่มต้น',
 
   //Error text form
+  'enter_title': 'กรุณากรอกชื่อหัวข้อ',
   'enter_info': 'กรอกข้อมูล',
   'enter_info_pls': 'กรุณากรอกข้อมูล',
   'enter_info_error': 'กรอกข้อมูลไม่สมบูรณ์',
@@ -252,10 +285,26 @@ const Map<String, String> th_TH = {
   'phone_10char': 'กรุณากรอกเบอร์โทรให้ครบ 10 ตัว',
   'select_type': 'กรุณาเลือกประเภท',
   'select_date': 'กรุณาเลือกวันที่',
+  'select_place': 'กรุณาเลือกสถานที่',
   'enter_access_count': 'กรุณาเพิ่มจำนวนสิทธิ์การใช้งาน',
 
   'enter_checklist': 'กรอกรายการตรวจ',
   'select_checklist': 'กรุณาเลือกรายการตรวจ',
   'select_event_pls': 'กรุณาเลือกเหตุการณ์',
-  'enter_desciption': 'เหตุการณ์ไม่ปกติ กรุณาเพิ่มรายละเอียด',
+  'ab_enter_description': 'เหตุการณ์ไม่ปกติ กรุณาเพิ่มรายละเอียด',
+  'enter_description': 'กรุณาเพิ่มรายละเอียด',
+
+  //Complaint staus
+  '1cp': 'รอการรับเรื่อง',
+  '2cp': 'รับเรื่องร้องเรียน',
+  '3cp': 'กำลังตรวจสอบ',
+  '4cp': 'ดำเนินการเสร็จสิ้น',
+  '5cp': 'ปิดเรื่องร้องเรียน',
+  '6cp': 'ยกเลิก',
+
+  //SOS staus
+  '1sos': 'รอการรับเรื่อง',
+  '2sos': 'กำลังตรวจสอบ',
+  '3sos': 'ดำเนินการเสร็จสิ้น',
+  '4sos': 'ยกเลิก',
 };
