@@ -1,13 +1,16 @@
 // ignore_for_file: unused_import
 
+import 'package:doormster/components/bottombar/navigation_ids.dart';
 import 'package:doormster/screen/management_service/management_service_page.dart';
 import 'package:flutter/material.dart';
 import 'package:doormster/screen/main_screen/home_page.dart';
 import 'package:doormster/screen/parcel_service/parcel_service_page.dart';
-import 'package:doormster/screen/qr_smart_access/qr_smart_home_page.dart';
-import 'package:doormster/screen/security_guard/security_guard_page.dart';
+import 'package:doormster/screen/qr_smart_access/smart_accress_menu.dart';
+// import 'package:doormster/screen/security_guard/security_guard_page.dart';
+import 'package:doormster/screen/security_guard/security_guard_menu.dart';
 import 'package:doormster/screen/visitor_service/visitor_service_page.dart';
 import 'package:doormster/components/bottombar/bottombar.dart';
+import 'package:get/get.dart';
 
 class Home_Menu extends StatefulWidget {
   const Home_Menu({Key? key});
@@ -20,34 +23,28 @@ class _Home_MenuState extends State<Home_Menu> {
   @override
   Widget build(BuildContext context) {
     return Navigator(
-      key: homeKey,
+      key: Keys.home,
       onGenerateRoute: (routeSettings) {
-        WidgetBuilder builder;
-        switch (routeSettings.name) {
-          case '/':
-            builder = (BuildContext context) => Home_Page();
-            break;
-          case '/qrsmart':
-            builder = (BuildContext context) => QRSmart_HomePage();
-            break;
-          case '/parcel':
-            builder = (BuildContext context) => Parcel_service();
-            break;
-          case '/management':
-            builder = (BuildContext context) => Management_Service();
-            break;
-          case '/security':
-            builder = (BuildContext context) => Security_Guard();
-            break;
-          case '/visitor':
-            builder = (BuildContext context) => Visitor_Service();
-            break;
-          default:
-            throw Exception('Invalid route: ${routeSettings.name}');
-        }
         return MaterialPageRoute(
           settings: routeSettings,
-          builder: builder,
+          builder: (context) {
+            switch (routeSettings.name) {
+              case '/':
+                return Home_Page();
+              case '/qrsmart':
+                return const Smart_Accress_Menu();
+              case '/parcel':
+                return Parcel_service();
+              case '/management':
+                return Management_Service();
+              case '/security':
+                return Security_Guard_Menu();
+              case '/visitor':
+                return Visitor_Service();
+              default:
+                throw Exception('Invalid route: ${routeSettings.name}');
+            }
+          },
         );
       },
     );
