@@ -1,5 +1,6 @@
 // ignore_for_file: use_key_in_widget_constructors, unused_import
 
+import 'package:doormster/components/bottombar/navigation_ids.dart';
 import 'package:doormster/screen/main_screen/notification_page.dart';
 import 'package:flutter/material.dart';
 import 'package:doormster/components/bottombar/bottombar.dart';
@@ -15,7 +16,7 @@ class _Notification_MenuState extends State<Notification_Menu> {
   @override
   Widget build(BuildContext context) {
     return Navigator(
-      key: notifyKey,
+      key: Keys.notify,
       onGenerateRoute: (routeSettings) {
         WidgetBuilder builder;
         switch (routeSettings.name) {
