@@ -1,5 +1,6 @@
 // ignore_for_file: avoid_unnecessary_containers, must_be_immutable, unused_import
 import 'package:doormster/components/bottombar/bottombar.dart';
+import 'package:doormster/components/bottombar/navigation_ids.dart';
 import 'package:doormster/screen/main_screen/profile_page.dart';
 import 'package:flutter/material.dart';
 
@@ -14,7 +15,7 @@ class _Profile_MenuState extends State<Profile_Menu> {
   @override
   Widget build(BuildContext context) {
     return Navigator(
-      key: profileKey,
+      key: Keys.profile,
       onGenerateRoute: (routeSettings) {
         WidgetBuilder builder;
         switch (routeSettings.name) {
