@@ -1,9 +1,14 @@
-// ignore_for_file: unrelated_type_equality_checks, prefer_const_constructors, use_build_context_synchronously, unused_import, prefer_typing_uninitialized_variables, camel_case_types
-
+// ignore_for_file: unrelated_type_equality_checks, prefer_const_constructors, use_build_context_synchronously, unused_import, prefer_typing_uninitialized_variables, camel_case_types, non_constant_identifier_names
+import 'package:dio/dio.dart' as dio;
 import 'package:doormster/components/actions/form_error_snackbar.dart';
+import 'package:doormster/components/dropdown/dropdown.dart';
 import 'package:doormster/components/text_form/text_form.dart';
+import 'package:doormster/components/text_form/text_form_number.dart';
 import 'package:doormster/components/text_form/text_form_validator.dart';
+import 'package:doormster/controller/add_images_controller.dart';
 import 'package:doormster/controller/profile_controller.dart';
+import 'package:doormster/models/main_models/get_profile.dart';
+import 'package:doormster/service/connected/ip_address.dart';
 import 'package:flutter/material.dart';
 import 'dart:developer';
 import 'dart:io';
@@ -16,7 +21,6 @@ import 'package:doormster/components/button/button.dart';
 import 'package:doormster/components/drawer/drawer.dart';
 import 'package:doormster/components/loading/loading.dart';
 import 'package:doormster/components/snackbar/snackbar.dart';
-import 'package:doormster/models/profile_model.dart';
 import 'package:doormster/service/connected/connect_api.dart';
 import 'package:doormster/service/permission/permission_camera.dart';
 import 'package:doormster/service/permission/permission_photos.dart';
@@ -34,45 +38,32 @@ class Profile_Update extends StatefulWidget {
 }
 
 class _Profile_UpdateState extends State<Profile_Update> {
-  final ProfileController controller = Get.find<ProfileController>();
+  final image = addImagesController.listImage;
 
   final _formkey = GlobalKey<FormState>();
   TextEditingController fname = TextEditingController();
   TextEditingController lname = TextEditingController();
+  TextEditingController phone = TextEditingController();
   TextEditingController email = TextEditingController();
+  TextEditingController sex = TextEditingController();
+  List<Map<String, String>> sexType = [
+    {'value': 'male', 'label': 'male'.tr},
+    {'value': 'female', 'label': 'female'.tr},
+  ];
   String? imageProfile;
 
-  List<Data> profileInfo = [];
-  var uuId;
+  Data profileInfo = Data();
+  String? userId;
+  String? dropdownValue;
 
   Future Controller() async {
-    fname = TextEditingController(text: profileInfo.single.firstName);
-    lname = TextEditingController(text: profileInfo.single.surName);
-    email = TextEditingController(text: profileInfo.single.email);
-    imageProfile = profileInfo.single.image;
-    uuId = profileInfo.single.userUuid;
-  }
-
-  final ImagePicker imgpicker = ImagePicker();
-  String? image;
-
-  Future<void> openImages(ImageSource TypeImage) async {
-    try {
-      // var pickedfiles = await imgpicker.pickMultiImage();
-      final XFile? photo = await imgpicker.pickImage(
-          source: TypeImage, maxHeight: 1080, maxWidth: 1080);
-      List<int> imageBytes = await photo!.readAsBytes();
-      var ImagesBase64 = convert.base64Encode(imageBytes);
-      if (photo != (null)) {
-        setState(() {
-          image = ImagesBase64;
-        });
-      } else {
-        print("No image is selected.");
-      }
-    } catch (e) {
-      print("error while picking file.");
-    }
+    fname = TextEditingController(text: profileInfo.firstName);
+    lname = TextEditingController(text: profileInfo.lastName);
+    phone = TextEditingController(text: profileInfo.phone);
+    email = TextEditingController(text: profileInfo.email);
+    sex = TextEditingController(text: profileInfo.sex!.tr);
+    imageProfile = profileInfo.profileImg;
+    userId = profileInfo.employeeNo;
   }
 
   @override
@@ -83,76 +74,77 @@ class _Profile_UpdateState extends State<Profile_Update> {
   }
 
   @override
+  void dispose() {
+    image.clear();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        Scaffold(
-          appBar: AppBar(
-            title: Text('edit_profile'.tr),
-          ),
-          floatingActionButtonLocation:
-              FloatingActionButtonLocation.centerFloat,
-          floatingActionButton: controller.loading.isTrue
-              ? Container()
-              : Buttons(
-                  title: 'save'.tr,
-                  press: () {
-                    if (_formkey.currentState!.validate()) {
-                      Map<String, dynamic> valuse = Map();
-                      valuse['uuid'] = uuId;
-                      valuse['first_name'] = fname.text;
-                      valuse['email'] = email.text;
-                      valuse['sur_name'] = lname.text;
-
-                      if (image != null) {
-                        valuse['image'] = image;
-                      } else {
-                        valuse['image'] = imageProfile;
+    return Obx(
+      () => Stack(
+        children: [
+          Scaffold(
+            appBar: AppBar(
+              title: Text('edit_profile'.tr),
+            ),
+            floatingActionButtonLocation:
+                FloatingActionButtonLocation.centerFloat,
+            floatingActionButton: connectApi.loading.isTrue
+                ? Container()
+                : Buttons(
+                    title: 'save'.tr,
+                    press: () async {
+                      List<dio.MultipartFile> files = [];
+                      for (int i = 0; i < image.length; i++) {
+                        files.add(await dio.MultipartFile.fromFile(
+                          image[i].path,
+                          filename: image[i].name,
+                        ));
                       }
-
-                      if (fname.text != profileInfo.single.firstName ||
-                          lname.text != profileInfo.single.surName ||
-                          email.text != profileInfo.single.email ||
-                          image != null) {
-                        controller.editProfile(valuse);
+                      if (_formkey.currentState!.validate()) {
+                        Map<String, dynamic> values = Map();
+                        values['employee_no'] = userId;
+                        values['first_name'] = fname.text;
+                        values['last_name'] = lname.text;
+                        values['phone'] = phone.text;
+                        values['email'] = email.text;
+                        values['sex'] = dropdownValue;
+                        if (files.isNotEmpty) {
+                          values['profileImg'] = files;
+                        }
+                        profileController.edit_Complaint(values);
                       } else {
-                        Get.back();
+                        form_error_snackbar();
                       }
-                    } else {
-                      form_error_snackbar();
-                    }
-                  }),
-          body: SingleChildScrollView(
-            physics: ClampingScrollPhysics(),
-            child: Form(
-              key: _formkey,
-              child: Container(
-                padding: EdgeInsets.fromLTRB(
-                    20, 20, 20, MediaQuery.of(context).size.height * 0.15),
-                child: Column(children: [
-                  Stack(
-                    children: [
-                      Card(
-                        elevation: 5,
-                        shape: CircleBorder(
-                            side: BorderSide(width: 3, color: Colors.white)),
-                        child: CircleAvatar(
-                            radius: 73,
-                            backgroundColor: Colors.grey.shade100,
-                            child: controller.loading.isTrue
-                                // || imageProfile == null
-                                ? Container(
-                                    child: Icon(Icons.person_rounded,
-                                        size: 140, color: Colors.grey),
-                                  )
-                                : image != null
+                    }),
+            body: SingleChildScrollView(
+              physics: ClampingScrollPhysics(),
+              child: Form(
+                key: _formkey,
+                child: Container(
+                  padding: EdgeInsets.fromLTRB(
+                      20, 20, 20, MediaQuery.of(context).size.height * 0.15),
+                  child: Column(children: [
+                    Stack(
+                      children: [
+                        Obx(() {
+                          return Card(
+                            elevation: 5,
+                            shape: CircleBorder(
+                                side:
+                                    BorderSide(width: 3, color: Colors.white)),
+                            child: CircleAvatar(
+                                radius: 73,
+                                backgroundColor: Colors.grey.shade100,
+                                child: image.isNotEmpty
                                     ? Container(
                                         decoration: BoxDecoration(
                                           shape: BoxShape.circle,
                                           image: DecorationImage(
                                               fit: BoxFit.cover,
-                                              image: MemoryImage(convert
-                                                  .base64Decode(image!))),
+                                              image: FileImage(
+                                                  File(image.first.path))),
                                         ),
                                       )
                                     : imageProfile != null
@@ -160,10 +152,18 @@ class _Profile_UpdateState extends State<Profile_Update> {
                                             decoration: BoxDecoration(
                                               shape: BoxShape.circle,
                                               image: DecorationImage(
-                                                  fit: BoxFit.cover,
-                                                  image: MemoryImage(
-                                                      convert.base64Decode(
-                                                          imageProfile!))),
+                                                fit: BoxFit.cover,
+                                                image: NetworkImage(
+                                                    imageDomain +
+                                                        imageProfile!),
+                                                onError: (exception,
+                                                        stackTrace) =>
+                                                    Icon(
+                                                        Icons
+                                                            .broken_image_rounded,
+                                                        size: 140,
+                                                        color: Colors.grey),
+                                              ),
                                             ),
                                           )
                                         : Container(
@@ -175,127 +175,109 @@ class _Profile_UpdateState extends State<Profile_Update> {
                                                       'assets/images/HIP Smart Community Icon-03.png')),
                                             ),
                                           )),
-                      ),
-                      Positioned(
-                        bottom: 0,
-                        right: 5,
-                        child: Card(
-                            elevation: 5,
-                            shape: CircleBorder(
-                                side: BorderSide(
-                                    color: Colors.white, width: 2.5)),
-                            child: CircleAvatar(
-                                radius: 21,
-                                backgroundColor: Theme.of(context).primaryColor,
-                                child: IconButton(
-                                  splashRadius: 20,
-                                  onPressed: () {
-                                    editImgProfile();
-                                  },
-                                  icon: Icon(
-                                    Icons.edit,
-                                    color: Colors.white,
-                                    size: 20,
-                                  ),
-                                ))),
-                      )
-                    ],
-                  ),
-                  SizedBox(
-                    height: 20,
-                  ),
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text('fname'.tr),
-                      Text_Form(
-                          controller: fname,
-                          title: 'fname'.tr,
-                          icon: Icons.person,
-                          error: 'enter_name_pls'.tr,
-                          TypeInput: TextInputType.name),
-                      SizedBox(height: 10),
-                      Text('lname'.tr),
-                      Text_Form(
-                        controller: lname,
-                        title: 'lname'.tr,
-                        icon: Icons.person_outline_rounded,
-                        error: 'enter_lname_pls'.tr,
-                        TypeInput: TextInputType.name,
-                      ),
-                      SizedBox(height: 10),
-                      Text('email'.tr),
-                      TextForm_validator(
-                          controller: email,
-                          title: 'email'.tr,
-                          icon: Icons.email,
-                          TypeInput: TextInputType.emailAddress,
+                          );
+                        }),
+                        Positioned(
+                          bottom: 0,
+                          right: 5,
+                          child: Card(
+                              elevation: 5,
+                              shape: CircleBorder(
+                                  side: BorderSide(
+                                      color: Colors.white, width: 2.5)),
+                              child: CircleAvatar(
+                                  radius: 21,
+                                  backgroundColor:
+                                      Theme.of(context).primaryColor,
+                                  child: IconButton(
+                                    splashRadius: 20,
+                                    onPressed: () {
+                                      addImagesController.selectType(
+                                          context, "change_picture".tr);
+                                    },
+                                    icon: Icon(
+                                      Icons.edit,
+                                      color: Colors.white,
+                                      size: 20,
+                                    ),
+                                  ))),
+                        )
+                      ],
+                    ),
+                    SizedBox(
+                      height: 20,
+                    ),
+                    Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text('fname'.tr),
+                        Text_Form(
+                            controller: fname,
+                            title: 'fname'.tr,
+                            icon: Icons.person,
+                            error: 'enter_name_pls'.tr,
+                            TypeInput: TextInputType.name),
+                        Text('lname'.tr),
+                        Text_Form(
+                          controller: lname,
+                          title: 'lname'.tr,
+                          icon: Icons.person_outline_rounded,
+                          error: 'enter_lname_pls'.tr,
+                          TypeInput: TextInputType.name,
+                        ),
+                        Text('phone_number'.tr),
+                        TextForm_Number(
+                          controller: phone,
+                          title: 'phone_number'.tr,
+                          icon: Icons.phone,
+                          type: TextInputType.name,
+                          maxLength: 10,
                           error: (values) {
                             if (values.isEmpty) {
-                              return 'enter_email_pls'.tr;
-                            } else if (!RegExp(
-                                    r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$')
-                                .hasMatch(values)) {
-                              return "email_error".tr;
+                              return 'enter_phone'.tr;
+                            } else if (values.length < 10) {
+                              return "phone_10char".tr;
                             } else {
                               return null;
                             }
-                          }),
-                    ],
-                  ),
-                ]),
+                          },
+                        ),
+                        Text('email'.tr),
+                        TextForm_validator(
+                            controller: email,
+                            title: 'email'.tr,
+                            icon: Icons.email,
+                            TypeInput: TextInputType.emailAddress,
+                            error: (values) {
+                              if (values.isEmpty) {
+                                return 'enter_email_pls'.tr;
+                              } else if (!RegExp(
+                                      r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$')
+                                  .hasMatch(values)) {
+                                return "email_error".tr;
+                              } else {
+                                return null;
+                              }
+                            }),
+                        Text('sex'.tr),
+                        Dropdown(
+                          title: 'sex'.tr,
+                          controller: sex,
+                          listItem: sexType.map((e) => e['label']!).toList(),
+                          onChanged: (value) {
+                            final selected = sexType
+                                .firstWhere((item) => item['label'] == value);
+                            dropdownValue = selected['value'];
+                          },
+                        )
+                      ],
+                    ),
+                  ]),
+                ),
               ),
             ),
           ),
-        ),
-        controller.loading.isTrue ? Loading() : Container()
-      ],
-    );
-  }
-
-  void editImgProfile() {
-    showDialog(
-      useRootNavigator: true,
-      barrierDismissible: true,
-      context: context,
-      builder: (_) => AlertDialog(
-        titlePadding: EdgeInsets.fromLTRB(30, 20, 20, 0),
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
-        title: Text(
-          'change_picture'.tr,
-          style: TextStyle(
-              fontSize: 16, letterSpacing: 0.5, fontWeight: FontWeight.bold),
-        ),
-        actions: [
-          ListTile(
-              title: Text('camera'.tr,
-                  style: TextStyle(fontSize: 15, letterSpacing: 0.5)),
-              onTap: () {
-                permissionCamere(context, () => openImages(ImageSource.camera));
-                Navigator.of(context, rootNavigator: true).pop();
-              }),
-          ListTile(
-            title: Text('photo'.tr,
-                style: TextStyle(fontSize: 15, letterSpacing: 0.5)),
-            onTap: () {
-              if (Platform.isIOS) {
-                permissionPhotos(context, openImages(ImageSource.gallery));
-              } else {
-                openImages(ImageSource.gallery);
-              }
-              Navigator.of(context, rootNavigator: true).pop();
-            },
-          ),
-          ListTile(
-            title: Text('cancel'.tr,
-                style: TextStyle(
-                  fontSize: 15,
-                  letterSpacing: 0.5,
-                )),
-            onTap: () {
-              Navigator.of(context, rootNavigator: true).pop();
-            },
-          ),
+          connectApi.loading.isTrue ? Loading() : Container()
         ],
       ),
     );
