@@ -1,7 +1,6 @@
 // ignore_for_file: use_build_context_synchronously, prefer_const_constructors, unused_import
 
 import 'dart:developer';
-
 import 'package:dio/dio.dart';
 import 'package:doormster/components/actions/disconnected_dialog.dart';
 import 'package:doormster/components/alertDialog/alert_dialog_onebutton_subtext.dart';
@@ -10,7 +9,8 @@ import 'package:doormster/components/list_null_opacity/logo_opacity.dart';
 import 'package:doormster/components/loading/loading.dart';
 import 'package:doormster/components/searchbar/search_from.dart';
 import 'package:doormster/components/text/text_double_colors.dart';
-import 'package:doormster/models/get_round.dart';
+import 'package:doormster/controller/security_controller/round_controller.dart';
+import 'package:doormster/models/secarity_models/get_round_all.dart';
 import 'package:doormster/screen/security_guard/scan_qr_check_page.dart';
 import 'package:doormster/service/connected/check_connected.dart';
 import 'package:doormster/service/connected/connect_api.dart';
@@ -24,8 +24,8 @@ import 'package:permission_handler/permission_handler.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class Round_Check extends StatefulWidget {
-  final checkpointId;
-  Round_Check({Key? key, this.checkpointId}) : super(key: key);
+  final logs;
+  Round_Check({Key? key, this.logs}) : super(key: key);
 
   @override
   State<Round_Check> createState() => _Round_CheckState();
@@ -33,16 +33,19 @@ class Round_Check extends StatefulWidget {
 
 class _Round_CheckState extends State<Round_Check> {
   TextEditingController fieldText = TextEditingController();
-  var companyId;
-  List<Data> listdata = [];
-  List<Data> listRound = [];
-  bool loading = false;
+  List<roundAll> listRound1 = roundController.listRound_All;
+  List<roundAll> listRound2 = roundController.listRound_All;
+
+  ScrollController scrollController = ScrollController();
+  int count = 6;
+  bool _hasMore = false;
+
   Color? containerColor;
   Color? textColor;
   Color? line;
+
   DateTime now = DateTime.now();
   int round = 0;
-  DateFormat format = DateFormat("HH:mm");
   DateTime? timeStart;
   DateTime? timeEnd;
 
@@ -81,60 +84,14 @@ class _Round_CheckState extends State<Round_Check> {
     }
   }
 
-  Future _getCheckRound(loadingTime) async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    companyId = prefs.getString('companyId');
-    now = DateTime.now();
-    print('companyId: ${companyId}');
-
-    try {
-      setState(() {
-        loading = true;
-      });
-
-      await Future.delayed(Duration(milliseconds: loadingTime));
-
-      //call api
-      var url = '${Connect_api().domain}/get/round/$companyId';
-      var response = await Dio().get(
-        url,
-        options: Options(headers: {
-          'Connect-type': 'application/json',
-          'Accept': 'application/json',
-        }),
-      );
-
-      if (response.statusCode == 200) {
-        getRound checklist = getRound.fromJson(response.data);
-        setState(() {
-          listdata = checklist.data!;
-          listRound = listdata;
-          loading = false;
-        });
-      }
-    } catch (error) {
-      print(error);
-      await Future.delayed(Duration(milliseconds: 500));
-      error_connected(() {
-        homeKey.currentState?.popUntil(ModalRoute.withName('/security'));
-        Navigator.of(context, rootNavigator: true).pop();
-      });
-      setState(() {
-        loading = false;
-      });
-    }
-  }
-
-  Future onGoBack(dynamic value) async {
-    setState(() {
-      _getCheckRound(0);
-    });
+  Future onGoBack() async {
+    roundController.get_roundAll(loadingTime: 0);
   }
 
   void _searchData(String text) {
     setState(() {
-      listdata = listRound.where((item) {
-        var name = item.roundName!;
+      listRound1 = listRound2.where((item) {
+        var name = item.inspectName!;
         return name.contains(text);
       }).toList();
     });
@@ -143,187 +100,228 @@ class _Round_CheckState extends State<Round_Check> {
   @override
   void initState() {
     super.initState();
-    _getCheckRound(300);
+    WidgetsBinding.instance.addPostFrameCallback((_) async {
+      roundController.get_roundAll(loadingTime: 100);
+    });
+    scrollController.addListener(
+      () async {
+        if (scrollController.position.pixels ==
+                scrollController.position.maxScrollExtent &&
+            !_hasMore) {
+          if (mounted) {
+            setState(() {
+              _hasMore = true;
+            });
+          }
+          await Future.delayed(const Duration(milliseconds: 800));
+          if (mounted) {
+            setState(() {
+              count = count + 4;
+              _hasMore = false;
+            });
+          }
+        }
+      },
+    );
+  }
+
+  @override
+  void dispose() {
+    scrollController.dispose();
+    super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     String date = DateFormat('y-MM-dd').format(now);
-    return Stack(
-      children: [
-        Scaffold(
-          appBar: AppBar(title: Text('list_round'.tr)),
-          body: loading
-              ? Container()
-              : Column(
-                  children: [
-                    listRound.length > 10
-                        ? Container(
-                            padding: const EdgeInsets.fromLTRB(20, 15, 20, 10),
-                            child: Search_From(
-                              title: 'search_round'.tr,
-                              fieldText: fieldText,
-                              clear: () {
-                                setState(() {
-                                  fieldText.clear();
-                                  listdata = listRound;
-                                });
-                              },
-                              changed: (value) {
-                                _searchData(value);
-                              },
-                            ),
-                          )
-                        : Container(),
-                    Expanded(
-                      child: listdata.isEmpty
-                          ? Logo_Opacity(title: 'no_data'.tr)
-                          : RefreshIndicator(
-                              onRefresh: () async {
-                                _getCheckRound(500);
-                              },
-                              child: ListView.builder(
-                                  padding: listdata.length > 10
-                                      ? const EdgeInsets.fromLTRB(20, 0, 20, 10)
-                                      : const EdgeInsets.symmetric(
-                                          vertical: 10, horizontal: 20),
-                                  itemCount: listdata.length,
-                                  itemBuilder: (context, index) {
-                                    timeStart = DateTime.parse(
-                                        '$date ${listdata[index].roundStart}');
-                                    timeEnd = DateTime.parse(
-                                        '$date ${listdata[index].roundEnd}');
-                                    _setColor();
-                                    return Card(
-                                      shape: RoundedRectangleBorder(
+    return Obx(
+      () => Scaffold(
+        appBar: AppBar(title: Text('list_round'.tr)),
+        body: connectApi.loading.isTrue
+            ? Loading()
+            : Column(
+                children: [
+                  listRound2.length > 10
+                      ? Container(
+                          padding: const EdgeInsets.fromLTRB(20, 15, 20, 10),
+                          child: Search_From(
+                            title: 'search_round'.tr,
+                            searchText: fieldText,
+                            clear: () {
+                              setState(() {
+                                fieldText.clear();
+                                listRound1 = listRound2;
+                              });
+                            },
+                            changed: (value) {
+                              _searchData(value);
+                            },
+                          ),
+                        )
+                      : Container(),
+                  Expanded(
+                    child: listRound1.isEmpty
+                        ? Logo_Opacity(title: 'no_data'.tr)
+                        : RefreshIndicator(
+                            onRefresh: () async {
+                              await roundController.get_roundAll(
+                                  loadingTime: 100);
+                            },
+                            child: ListView.builder(
+                                padding: listRound1.length > 10
+                                    ? const EdgeInsets.fromLTRB(20, 0, 20, 10)
+                                    : const EdgeInsets.symmetric(
+                                        vertical: 10, horizontal: 20),
+                                controller: scrollController,
+                                itemCount: count < listRound1.length
+                                    ? count + (_hasMore ? 1 : 0)
+                                    : listRound1.length,
+                                itemBuilder: (context, index) {
+                                  if (index >= count) {
+                                    return const Padding(
+                                      padding: EdgeInsets.only(top: 20),
+                                      child: Center(
+                                          child: CircularProgressIndicator()),
+                                    );
+                                  }
+                                  timeStart = DateTime.parse(
+                                      '$date ${listRound1[index].startDate}');
+                                  timeEnd = DateTime.parse(
+                                      '$date ${listRound1[index].endDate}');
+                                  _setColor();
+                                  return Card(
+                                    shape: RoundedRectangleBorder(
+                                        borderRadius:
+                                            BorderRadius.circular(10)),
+                                    margin: EdgeInsets.symmetric(vertical: 5),
+                                    elevation: 10,
+                                    child: Container(
+                                      decoration: BoxDecoration(
+                                          color: containerColor,
                                           borderRadius:
                                               BorderRadius.circular(10)),
-                                      margin: EdgeInsets.symmetric(vertical: 5),
-                                      elevation: 10,
-                                      child: Container(
-                                        decoration: BoxDecoration(
-                                            color: containerColor,
-                                            borderRadius:
-                                                BorderRadius.circular(10)),
-                                        padding: EdgeInsets.all(10),
-                                        child: Column(
-                                          crossAxisAlignment:
-                                              CrossAxisAlignment.center,
-                                          children: [
-                                            round == 1
-                                                ? Column(
-                                                    children: [
-                                                      Text(
-                                                        'need_check'.tr,
-                                                        style: TextStyle(
-                                                            color: textColor),
-                                                      ),
-                                                      Divider(
-                                                          height: 15,
-                                                          thickness: 1.5,
-                                                          color: line),
-                                                    ],
-                                                  )
-                                                : Container(),
-                                            Row(
+                                      padding: EdgeInsets.all(10),
+                                      child: Column(
+                                        crossAxisAlignment:
+                                            CrossAxisAlignment.center,
+                                        children: [
+                                          round == 1
+                                              ? Column(
+                                                  children: [
+                                                    Text(
+                                                      'need_check'.tr,
+                                                      style: TextStyle(
+                                                          color: textColor),
+                                                    ),
+                                                    Divider(
+                                                        height: 15,
+                                                        thickness: 1.5,
+                                                        color: line),
+                                                  ],
+                                                )
+                                              : Container(),
+                                          Row(
+                                            mainAxisAlignment:
+                                                MainAxisAlignment.spaceBetween,
+                                            children: [
+                                              Expanded(
+                                                  child: Text(
+                                                '${'round'.tr} : ${listRound1[index].inspectName}',
+                                                style:
+                                                    TextStyle(color: textColor),
+                                              )),
+                                              round == 1
+                                                  ? button(
+                                                      'checkIn'.tr,
+                                                      Theme.of(context)
+                                                          .cardTheme
+                                                          .color, () {
+                                                      listRound1[index]
+                                                          .inspectDetail;
+                                                      permissionCamere(
+                                                          context,
+                                                          () => permissionLocation(
+                                                              context,
+                                                              () => checkInternet(
+                                                                  page: ScanQR_Check(
+                                                                      name:
+                                                                          'check',
+                                                                      roundId:
+                                                                          listRound1[index]
+                                                                              .inspectId!,
+                                                                      roundName:
+                                                                          listRound1[index]
+                                                                              .inspectName!,
+                                                                      roundStart:
+                                                                          listRound1[index]
+                                                                              .startDate!,
+                                                                      roundEnd:
+                                                                          listRound1[index]
+                                                                              .endDate!,
+                                                                      page:
+                                                                          'round',
+                                                                      inspectDetail:
+                                                                          listRound1[index].inspectDetail ??
+                                                                              [],
+                                                                      logs: widget
+                                                                          .logs),
+                                                                  onGoBack:
+                                                                      onGoBack)));
+                                                      // : dialogOnebutton_Subtitle(
+                                                      //     context,
+                                                      //     'ไม่สามารถตรวจได้',
+                                                      //     'เลยเวลาเดินตรวจรอบนี้แล้ว',
+                                                      //     Icons
+                                                      //         .warning_amber_rounded,
+                                                      //     Colors.orange,
+                                                      //     'ตกลง', () {
+                                                      //     Navigator.of(
+                                                      //             context,
+                                                      //             rootNavigator:
+                                                      //                 true)
+                                                      //         .pop();
+                                                      //   }, false, false);
+                                                    })
+                                                  : Container()
+                                            ],
+                                          ),
+                                          Divider(
+                                              height: 15,
+                                              thickness: 1.5,
+                                              color: line),
+                                          IntrinsicHeight(
+                                            child: Row(
                                               mainAxisAlignment:
                                                   MainAxisAlignment
                                                       .spaceBetween,
                                               children: [
-                                                Expanded(
-                                                    child: Text(
-                                                  '${'round'.tr} : ${listdata[index].roundName}',
+                                                Text(
+                                                  '${'start'.tr} : ${listRound1[index].startDate!.substring(0, 5)}',
                                                   style: TextStyle(
                                                       color: textColor),
-                                                )),
-                                                round == 1
-                                                    ? button(
-                                                        'checkIn'.tr,
-                                                        Theme.of(context)
-                                                            .cardTheme
-                                                            .color, () {
-                                                        permissionCamere(
-                                                            context,
-                                                            () => permissionLocation(
-                                                                context,
-                                                                () => checkInternetOnGoBack(
-                                                                    context,
-                                                                    ScanQR_Check(
-                                                                        name:
-                                                                            'check',
-                                                                        roundId:
-                                                                            listdata[index]
-                                                                                .roundUuid!,
-                                                                        roundName:
-                                                                            listdata[index]
-                                                                                .roundName!,
-                                                                        roundStart:
-                                                                            listdata[index]
-                                                                                .roundStart!,
-                                                                        roundEnd:
-                                                                            listdata[index]
-                                                                                .roundEnd!,
-                                                                        checkpointId:
-                                                                            widget.checkpointId),
-                                                                    true,
-                                                                    onGoBack)));
-                                                        // : dialogOnebutton_Subtitle(
-                                                        //     context,
-                                                        //     'ไม่สามารถตรวจได้',
-                                                        //     'เลยเวลาเดินตรวจรอบนี้แล้ว',
-                                                        //     Icons
-                                                        //         .warning_amber_rounded,
-                                                        //     Colors.orange,
-                                                        //     'ตกลง', () {
-                                                        //     Navigator.of(
-                                                        //             context,
-                                                        //             rootNavigator:
-                                                        //                 true)
-                                                        //         .pop();
-                                                        //   }, false, false);
-                                                      })
-                                                    : Container()
+                                                ),
+                                                VerticalDivider(
+                                                    thickness: 1.5,
+                                                    color: line,
+                                                    width: 1),
+                                                Text(
+                                                  '${'end'.tr} : ${listRound1[index].endDate!.substring(0, 5)}',
+                                                  style: TextStyle(
+                                                      color: textColor),
+                                                ),
                                               ],
                                             ),
-                                            Divider(
-                                                height: 15,
-                                                thickness: 1.5,
-                                                color: line),
-                                            IntrinsicHeight(
-                                              child: Row(
-                                                mainAxisAlignment:
-                                                    MainAxisAlignment
-                                                        .spaceBetween,
-                                                children: [
-                                                  Text(
-                                                    '${'start'.tr} : ${listdata[index].roundStart}',
-                                                    style: TextStyle(
-                                                        color: textColor),
-                                                  ),
-                                                  VerticalDivider(
-                                                      thickness: 1.5,
-                                                      color: line,
-                                                      width: 1),
-                                                  Text(
-                                                    '${'end'.tr} : ${listdata[index].roundEnd}',
-                                                    style: TextStyle(
-                                                        color: textColor),
-                                                  ),
-                                                ],
-                                              ),
-                                            )
-                                          ],
-                                        ),
+                                          )
+                                        ],
                                       ),
-                                    );
-                                  }),
-                            ),
-                    ),
-                  ],
-                ),
-        ),
-        loading ? Loading() : Container()
-      ],
+                                    ),
+                                  );
+                                }),
+                          ),
+                  ),
+                ],
+              ),
+      ),
     );
   }
 
@@ -332,10 +330,10 @@ class _Round_CheckState extends State<Round_Check> {
       style: TextButton.styleFrom(
         shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
         elevation: 10,
-        primary: Colors.white,
         backgroundColor: color,
         padding: EdgeInsets.symmetric(vertical: 5, horizontal: 10),
       ),
+      onPressed: press,
       child: Text(
         name,
         style: TextStyle(
@@ -343,7 +341,6 @@ class _Round_CheckState extends State<Round_Check> {
             color: Theme.of(context).dividerColor,
             fontWeight: FontWeight.normal),
       ),
-      onPressed: press,
     );
   }
 }
