@@ -1,10 +1,12 @@
 import 'package:doormster/components/girdManu/grid_menu.dart';
+import 'package:doormster/components/list_null_opacity/logo_opacity.dart';
 import 'package:doormster/screen/visitor_service/scan_idcard_page.dart';
+import 'package:doormster/screen/visitor_service/visitor_idcard_page.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 
 class Visitor_Service extends StatefulWidget {
   Visitor_Service({Key? key});
-  static const String route = '/visitor';
 
   @override
   State<Visitor_Service> createState() => _Visitor_ServiceState();
@@ -17,7 +19,9 @@ class _Visitor_ServiceState extends State<Visitor_Service> {
       appBar: AppBar(
         title: Text('Visitor'),
       ),
-      body: Stack(
+      body:
+          // Logo_Opacity(title: 'no_data'.tr),
+          Stack(
         alignment: Alignment.center,
         children: [
           GridView.count(
@@ -34,7 +38,7 @@ class _Visitor_ServiceState extends State<Visitor_Service> {
                   press: () {
                     Navigator.of(context, rootNavigator: true)
                         .push(MaterialPageRoute(
-                      builder: (context) => Scan_IDCard(),
+                      builder: (context) => Visitor_IDCard(),
                     ));
                   },
                   icon: Icons.document_scanner_outlined),
