@@ -1,23 +1,28 @@
 import 'package:connectivity_plus/connectivity_plus.dart';
+import 'package:doormster/components/bottombar/navigation_ids.dart';
 import 'package:doormster/components/snackbar/snackbar.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-Future<void> checkInternet(context, page, rootNavi) async {
+Future<void> checkInternet(
+    {required Widget page, navigationId, onGoBack}) async {
   var result = await Connectivity().checkConnectivity();
-  print(result);
   if (result == ConnectivityResult.none) {
     snackbar(
         Colors.orange, 'connect_internet_pls'.tr, Icons.warning_amber_rounded);
     print('not connected');
   } else {
-    Navigator.of(context, rootNavigator: rootNavi)
-        .push(MaterialPageRoute(builder: (context) => page));
+    if (onGoBack != null) {
+      Get.to(() => page, id: navigationId)!.then(
+        (_) => onGoBack(),
+      );
+    } else {
+      Get.to(() => page, id: navigationId);
+    }
   }
 }
 
-Future<void> checkInternetOnGoBack(
-    context, Widget page, rootNavi, onGoBack) async {
+Future<void> checkInternetName({required page, onGoBack}) async {
   var result = await Connectivity().checkConnectivity();
   print(result);
   if (result == ConnectivityResult.none) {
@@ -25,35 +30,12 @@ Future<void> checkInternetOnGoBack(
         Colors.orange, 'connect_internet_pls'.tr, Icons.warning_amber_rounded);
     print('not connected');
   } else {
-    Navigator.of(context, rootNavigator: rootNavi)
-        .push(MaterialPageRoute(builder: (context) => page))
-        .then(onGoBack);
-  }
-}
-
-Future<void> checkInternetName(context, page, rootNavi) async {
-  var result = await Connectivity().checkConnectivity();
-  print(result);
-  if (result == ConnectivityResult.none) {
-    snackbar(
-        Colors.orange, 'connect_internet_pls'.tr, Icons.warning_amber_rounded);
-    print('not connected');
-  } else {
-    Navigator.of(context, rootNavigator: rootNavi).pushNamed(page);
-  }
-}
-
-Future<void> checkInternetNameOnGoBack(
-    context, page, rootNavi, onGoBack) async {
-  var result = await Connectivity().checkConnectivity();
-  print(result);
-  if (result == ConnectivityResult.none) {
-    snackbar(
-        Colors.orange, 'connect_internet_pls'.tr, Icons.warning_amber_rounded);
-    print('not connected');
-  } else {
-    Navigator.of(context, rootNavigator: rootNavi)
-        .pushNamed(page)
-        .then((onGoBack));
+    if (onGoBack != null) {
+      Get.toNamed(page, id: NavigationIds.home)?.then(
+        (_) => onGoBack(),
+      );
+    } else {
+      Get.toNamed(page, id: NavigationIds.home);
+    }
   }
 }
