@@ -5,6 +5,7 @@ import 'dart:io';
 
 import 'package:doormster/components/bottombar/bottom_controller.dart';
 import 'package:doormster/components/bottombar/bottombar.dart';
+import 'package:doormster/components/bottombar/navigation_ids.dart';
 import 'package:doormster/routes/menu/notification_menu.dart';
 import 'package:doormster/screen/main_screen/notification_page.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
@@ -14,17 +15,46 @@ import 'package:get/get.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 Future<void> handleBackgroundMessage(RemoteMessage message) async {
+  final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
+      FlutterLocalNotificationsPlugin();
   log("BackgroundMessage ");
-  if (message != null) {
-    notifyKey.currentState?.push(
-      GetPageRoute(
-          page: () => Notification_Page(
-                title: message.notification?.title,
-                body: message.notification?.body,
-              ),
-          transitionDuration: Duration.zero),
-    );
-  }
+  BigTextStyleInformation bigTextStyleInformation = BigTextStyleInformation(
+    message.notification!.body.toString(),
+    htmlFormatBigText: true,
+    contentTitle: message.notification!.title.toString(),
+    htmlFormatContentTitle: true,
+  );
+  AndroidNotificationDetails androidChannel = AndroidNotificationDetails(
+    'high_importance_channel', // id
+    'High Importance Notifications', // title
+    importance: Importance.max,
+    priority: Priority.high,
+    ticker: 'ticker',
+    icon: 'icon_app',
+    channelShowBadge: true,
+    styleInformation: bigTextStyleInformation,
+    // color: Theme.of(context).primaryColor,
+    // largeIcon: DrawableResourceAndroidBitmap('circle_icon'),
+  );
+  DarwinNotificationDetails iosChannel = const DarwinNotificationDetails(
+    presentAlert: true,
+    presentBadge: true,
+    presentSound: true,
+  );
+  NotificationDetails platformChannel =
+      NotificationDetails(android: androidChannel, iOS: iosChannel);
+
+  await flutterLocalNotificationsPlugin.show(1, message.notification?.title,
+      message.notification?.body, platformChannel,
+      payload: message.data['body']);
+  Keys.notify?.currentState?.push(
+    GetPageRoute(
+        page: () => Notification_Page(
+              title: message.notification?.title,
+              body: message.notification?.body,
+            ),
+        transitionDuration: Duration.zero),
+  );
 }
 
 void notificationTapBackground(NotificationResponse notificationResponse) {
@@ -69,16 +99,14 @@ class NotificationService {
         payload: message.data['body']);
     log("ForegroundMessage");
 
-    // if (message != null) {
-    //   notifyKey.currentState?.pushReplacement(
-    //     GetPageRoute(
-    //         page: () => Notification_Page(
-    //               title: message.notification?.title,
-    //               body: message.notification?.body,
-    //             ),
-    //         transitionDuration: Duration.zero),
-    //   );
-    // }
+    Keys.notify?.currentState?.pushReplacement(
+      GetPageRoute(
+          page: () => Notification_Page(
+                title: message.notification?.title,
+                body: message.notification?.body,
+              ),
+          transitionDuration: Duration.zero),
+    );
   }
 
   Future<void> notification() async {
