@@ -13,36 +13,57 @@ Future<void> permissionLocation(context, action) async {
   print("location : $location");
 
   if (Platform.isAndroid) {
-    print('divices : android');
     if (permission.isGranted) {
       await action();
     } else {
-      dialogOnebutton_Subtitle('allow_access'.tr, 'need_access_location'.tr,
-          Icons.warning_amber_rounded, Colors.orange, 'ok'.tr, () {
-        openAppSettings();
-
-        Navigator.of(context, rootNavigator: true).pop();
-      }, true, true);
+      dialogOnebutton_Subtitle(
+          title: 'allow_access'.tr,
+          subtitle: 'need_access_location'.tr,
+          icon: Icons.warning_amber_rounded,
+          colorIcon: Colors.orange,
+          textButton: 'ok'.tr,
+          press: () {
+            openAppSettings();
+            Navigator.of(context, rootNavigator: true).pop();
+          },
+          click: true,
+          backBtn: true,
+          willpop: true);
     }
   } else if (Platform.isIOS) {
-    print("divices : apple");
     if (permission.isGranted && location) {
       await action();
     } else if (!location) {
-      dialogOnebutton_Subtitle('access_location'.tr, 'turn_on_location'.tr,
-          Icons.warning_amber_rounded, Colors.orange, 'ok'.tr, () async {
-        await launch('App-Prefs:LOCATION_SERVICES');
+      dialogOnebutton_Subtitle(
+          title: 'access_location'.tr,
+          subtitle: 'turn_on_location'.tr,
+          icon: Icons.warning_amber_rounded,
+          colorIcon: Colors.orange,
+          textButton: 'ok'.tr,
+          press: () async {
+            await launch('App-Prefs:LOCATION_SERVICES');
 
-        Navigator.of(context, rootNavigator: true).pop();
-      }, true, true);
+            Navigator.of(context, rootNavigator: true).pop();
+          },
+          click: true,
+          backBtn: true,
+          willpop: true);
     } else {
-      dialogOnebutton_Subtitle('allow_access'.tr, 'need_access_location'.tr,
-          Icons.warning_amber_rounded, Colors.orange, 'ok'.tr, () {
-        print('setting');
-        openAppSettings();
+      dialogOnebutton_Subtitle(
+          title: 'allow_access'.tr,
+          subtitle: 'need_access_location'.tr,
+          icon: Icons.warning_amber_rounded,
+          colorIcon: Colors.orange,
+          textButton: 'ok'.tr,
+          press: () {
+            print('setting');
+            openAppSettings();
 
-        Navigator.of(context, rootNavigator: true).pop();
-      }, true, true);
+            Navigator.of(context, rootNavigator: true).pop();
+          },
+          click: true,
+          backBtn: true,
+          willpop: true);
     }
   }
 }
