@@ -11,11 +11,19 @@ Future<void> permissionPhotos(context, action) async {
   if (permission.isGranted || permission.isLimited) {
     action;
   } else {
-    dialogOnebutton_Subtitle('allow_access'.tr, 'need_access_picture'.tr,
-        Icons.warning_amber_rounded, Colors.orange, 'ok'.tr, () {
-      Navigator.of(context, rootNavigator: true).pop();
-      openAppSettings();
-    }, true, true);
+    dialogOnebutton_Subtitle(
+        title: 'allow_access'.tr,
+        subtitle: 'need_access_picture'.tr,
+        icon: Icons.warning_amber_rounded,
+        colorIcon: Colors.orange,
+        textButton: 'ok'.tr,
+        press: () {
+          Navigator.of(context, rootNavigator: true).pop();
+          openAppSettings();
+        },
+        click: true,
+        backBtn: true,
+        willpop: true);
   }
 }
 
@@ -26,11 +34,19 @@ Future<void> permissionAddPhotos(context, action) async {
     if (permission.isGranted || permission.isLimited) {
       await action();
     } else {
-      dialogOnebutton_Subtitle('allow_access'.tr, 'need_access_picture'.tr,
-          Icons.warning_amber_rounded, Colors.orange, 'ok'.tr, () {
-        Navigator.of(context, rootNavigator: true).pop();
-        openAppSettings();
-      }, true, true);
+      dialogOnebutton_Subtitle(
+          title: 'allow_access'.tr,
+          subtitle: 'need_access_picture'.tr,
+          icon: Icons.warning_amber_rounded,
+          colorIcon: Colors.orange,
+          textButton: 'ok'.tr,
+          press: () {
+            Navigator.of(context, rootNavigator: true).pop();
+            openAppSettings();
+          },
+          click: true,
+          backBtn: true,
+          willpop: true);
     }
   } else {
     await action();
