@@ -2,12 +2,10 @@ import 'package:flutter/material.dart';
 
 Widget myScrollScreen(Widget child) {
   return ScrollConfiguration(
-    // scroll listview screen การเลื่อนไลด์หน้าแอพ
-    behavior: ScrollBehavior().copyWith(
-        // Set the default scroll physics here
-        physics: BouncingScrollPhysics()
-        // Platform.isIOS ? BouncingScrollPhysics() : ClampingScrollPhysics(),
-        ),
+    // scroll listview screen
+    behavior: const ScrollBehavior().copyWith(
+        physics: const BouncingScrollPhysics(
+            parent: AlwaysScrollableScrollPhysics())),
     child: child,
   );
 }
