@@ -10,6 +10,6 @@ BottomNavigationBarThemeData Dark_BottomBar = BottomNavigationBarThemeData(
     selectedItemColor: Colors.white,
     unselectedItemColor: Colors.grey,
     selectedIconTheme: IconThemeData(size: 30),
-    unselectedIconTheme: IconThemeData(size: 25),
+    unselectedIconTheme: IconThemeData(size: 30),
     selectedLabelStyle: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-    unselectedLabelStyle: TextStyle(fontSize: 14));
+    unselectedLabelStyle: TextStyle(fontSize: 16));
